Use CachedItemFileProvider.kt
